Rename `_ALIGNBYTES_` to `CUBISM_ALIGNMENT`.
Not sure about this, but now Cubism automatically defines
`CUBISM_ALIGNMENT` if it is not already defined.
That means that if the app intends to define the value,
it must define it before including _any_ Cubism header.

An alternative would be to put these ifdefs to Matrix3D.h (the only file
using the constant), but probably better to be strict.