HDF5: Rejected MeshGridFix
The code would not be correct because Grid::getH() returns the (legacy)
uniform grid spacing h_gridpoint which is identical for all dimensions.
Testing for MeshMap::uniform() means uniform in the corresponding
dimension but it is possible that hx != hy != hz in the most general
case.  In this case Grid::getH() does not work.